+++ src/main/scala/com/scienaptic/jobs/core/npd/pc/monthly/transformations/USTransformer.scala
@@ -12,7 +12,7 @@ object USTransformer {
   def withAllTransformations(df : DataFrame) = {
 
     val cleanUpDollers = (str : String) => {
-      str.replace("$","").replace(",","").toInt
+      str.replace("$","").replace(",","").toDouble
     }
 
     def cleanDollersUDF = udf(cleanUpDollers)
